+++ examples/simple_mbtiles.py
@@ -9,18 +9,18 @@ is not the case on every platform at 1.8.1, but we're going to fix it.
 """
 
 import sys
+
 from kivy.base import runTouchApp
 
-if __name__ == '__main__' and __package__ is None:
-    from os import path
-    sys.path.append(path.dirname(path.dirname(path.abspath(__file__))))
 from kivy_garden.mapview import MapView
 from kivy_garden.mapview.mbtsource import MBTilesMapSource
 
-
 source = MBTilesMapSource(sys.argv[1])
-runTouchApp(MapView(
-    map_source=source,
-    lat=source.default_lat,
-    lon=source.default_lon,
-    zoom=source.default_zoom))
+runTouchApp(
+    MapView(
+        map_source=source,
+        lat=source.default_lat,
+        lon=source.default_lon,
+        zoom=source.default_zoom,
+    )
+)