+++ kivy_garden/mapview/__init__.py
@@ -7,26 +7,23 @@ MapView
 
 MapView is a Kivy widget that display maps.
 """
-
-__all__ = ["Coordinate", "Bbox", "MapView", "MapSource", "MapMarker",
-           "MapLayer", "MarkerMapLayer", "MapMarkerPopup"]
-__version__ = "0.2"
-
-MIN_LATITUDE = -90.
-MAX_LATITUDE = 90.
-MIN_LONGITUDE = -180.
-MAX_LONGITUDE = 180.
-CACHE_DIR = "cache"
-
-try:
-    # fix if used within garden
-    import sys
-    sys.modules['mapview'] = sys.modules['kivy.garden.mapview.mapview']
-    del sys
-except KeyError:
-    pass
-
-from kivy_garden.mapview.types import Coordinate, Bbox
 from kivy_garden.mapview.source import MapSource
-from kivy_garden.mapview.view import MapView, MapMarker, MapLayer, MarkerMapLayer, \
-    MapMarkerPopup
+from kivy_garden.mapview.types import Bbox, Coordinate
+from kivy_garden.mapview.view import (
+    MapLayer,
+    MapMarker,
+    MapMarkerPopup,
+    MapView,
+    MarkerMapLayer,
+)
+
+__all__ = [
+    "Coordinate",
+    "Bbox",
+    "MapView",
+    "MapSource",
+    "MapMarker",
+    "MapLayer",
+    "MarkerMapLayer",
+    "MapMarkerPopup",
+]