+++ kivy_garden/mapview/view.py
@@ -1,32 +1,45 @@
 # coding=utf-8
 
-__all__ = ["MapView", "MapMarker", "MapMarkerPopup", "MapLayer",
-           "MarkerMapLayer"]
+__all__ = ["MapView", "MapMarker", "MapMarkerPopup", "MapLayer", "MarkerMapLayer"]
 
-from os.path import join, dirname
+import webbrowser
+from itertools import takewhile
+from math import ceil
+from os.path import dirname, join
+
 from kivy.clock import Clock
-from kivy.metrics import dp
-from kivy.uix.widget import Widget
-from kivy.uix.label import Label
-from kivy.uix.image import Image
-from kivy.uix.scatter import Scatter
-from kivy.uix.behaviors import ButtonBehavior
-from kivy.properties import NumericProperty, ObjectProperty, ListProperty, \
-    AliasProperty, BooleanProperty, StringProperty
+from kivy.compat import string_types
 from kivy.graphics import Canvas, Color, Rectangle
 from kivy.graphics.transformation import Matrix
 from kivy.lang import Builder
-from kivy.compat import string_types
-from math import ceil
-from kivy_garden.mapview import MIN_LONGITUDE, MAX_LONGITUDE, MIN_LATITUDE, MAX_LATITUDE, \
-    CACHE_DIR, Coordinate, Bbox
+from kivy.metrics import dp
+from kivy.properties import (
+    AliasProperty,
+    BooleanProperty,
+    ListProperty,
+    NumericProperty,
+    ObjectProperty,
+    StringProperty,
+)
+from kivy.uix.behaviors import ButtonBehavior
+from kivy.uix.image import Image
+from kivy.uix.label import Label
+from kivy.uix.scatter import Scatter
+from kivy.uix.widget import Widget
+
+from kivy_garden.mapview import Bbox, Coordinate
+from kivy_garden.mapview.constants import (
+    CACHE_DIR,
+    MAX_LATITUDE,
+    MAX_LONGITUDE,
+    MIN_LATITUDE,
+    MIN_LONGITUDE,
+)
 from kivy_garden.mapview.source import MapSource
 from kivy_garden.mapview.utils import clamp
-from itertools import takewhile
 
-import webbrowser
-
-Builder.load_string("""
+Builder.load_string(
+    """
 <MapMarker>:
     size_hint: None, None
     source: root.source
@@ -81,7 +94,8 @@ Builder.load_string("""
         center_x: root.center_x
         size: root.popup_size
 
-""")
+"""
+)
 
 
 class ClickableLabel(Label):
@@ -91,7 +105,7 @@ class ClickableLabel(Label):
 
 class Tile(Rectangle):
     def __init__(self, *args, **kwargs):
-        super(Tile, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
         self.cache_dir = kwargs.get('cache_dir', CACHE_DIR)
 
     @property
@@ -100,7 +114,8 @@ class Tile(Rectangle):
         fn = map_source.cache_fmt.format(
             image_ext=map_source.image_ext,
             cache_key=map_source.cache_key,
-            **self.__dict__)
+            **self.__dict__
+        )
         return join(self.cache_dir, fn)
 
     def set_source(self, cache_fn):
@@ -152,7 +167,7 @@ class MapMarkerPopup(MapMarker):
         if not self.placeholder:
             self.placeholder = widget
             if self.is_open:
-                super(MapMarkerPopup, self).add_widget(self.placeholder)
+                super().add_widget(self.placeholder)
         else:
             self.placeholder.add_widget(widget)
 
@@ -160,7 +175,7 @@ class MapMarkerPopup(MapMarker):
         if widget is not self.placeholder:
             self.placeholder.remove_widget(widget)
         else:
-            super(MapMarkerPopup, self).remove_widget(widget)
+            super().remove_widget(widget)
 
     def on_is_open(self, *args):
         self.refresh_open_status()
@@ -170,15 +185,16 @@ class MapMarkerPopup(MapMarker):
 
     def refresh_open_status(self):
         if not self.is_open and self.placeholder.parent:
-            super(MapMarkerPopup, self).remove_widget(self.placeholder)
+            super().remove_widget(self.placeholder)
         elif self.is_open and not self.placeholder.parent:
-            super(MapMarkerPopup, self).add_widget(self.placeholder)
+            super().add_widget(self.placeholder)
 
 
 class MapLayer(Widget):
     """A map layer, that is repositionned everytime the :class:`MapView` is
     moved.
     """
+
     viewport_x = NumericProperty(0)
     viewport_y = NumericProperty(0)
 
@@ -197,22 +213,22 @@ class MapLayer(Widget):
 class MarkerMapLayer(MapLayer):
     """A map layer for :class:`MapMarker`
     """
+
     order_marker_by_latitude = BooleanProperty(True)
 
     def __init__(self, **kwargs):
         self.markers = []
-        super(MarkerMapLayer, self).__init__(**kwargs)
+        super().__init__(**kwargs)
 
     def insert_marker(self, marker, **kwargs):
         if self.order_marker_by_latitude:
-            before = list(takewhile(
-                lambda i_m: i_m[1].lat < marker.lat,
-                enumerate(self.children)
-            ))
+            before = list(
+                takewhile(lambda i_m: i_m[1].lat < marker.lat, enumerate(self.children))
+            )
             if before:
                 kwargs['index'] = before[-1][0] + 1
 
-        super(MarkerMapLayer, self).add_widget(marker, **kwargs)
+        super().add_widget(marker, **kwargs)
 
     def add_widget(self, marker):
         marker._layer = self
@@ -223,7 +239,7 @@ class MarkerMapLayer(MapLayer):
         marker._layer = None
         if marker in self.markers:
             self.markers.remove(marker)
-        super(MarkerMapLayer, self).remove_widget(marker)
+        super().remove_widget(marker)
 
     def reposition(self):
         if not self.markers:
@@ -241,7 +257,7 @@ class MarkerMapLayer(MapLayer):
                 if not marker.parent:
                     self.insert_marker(marker)
             else:
-                super(MarkerMapLayer, self).remove_widget(marker)
+                super().remove_widget(marker)
 
     def set_marker_position(self, mapview, marker):
         x, y = mapview.get_window_xy_from(marker.lat, marker.lon, mapview.zoom)
@@ -256,11 +272,10 @@ class MarkerMapLayer(MapLayer):
 class MapViewScatter(Scatter):
     # internal
     def on_transform(self, *args):
-        super(MapViewScatter, self).on_transform(*args)
+        super().on_transform(*args)
         self.parent.on_transform(self.transform)
 
     def collide_point(self, x, y):
-        # print "collide_point", x, y
         return True
 
 
@@ -309,11 +324,11 @@ class MapView(Widget):
 
     delta_x = NumericProperty(0)
     delta_y = NumericProperty(0)
-    background_color = ListProperty([181 / 255., 208 / 255., 208 / 255., 1])
+    background_color = ListProperty([181 / 255.0, 208 / 255.0, 208 / 255.0, 1])
     cache_dir = StringProperty(CACHE_DIR)
     _zoom = NumericProperty(0)
     _pause = BooleanProperty(False)
-    _scale = 1.
+    _scale = 1.0
     _disabled_count = 0
 
     __events__ = ["on_map_relocated"]
@@ -337,8 +352,7 @@ class MapView(Widget):
         top/right (lat2, lon2).
         """
         x1, y1 = self.to_local(0 - margin, 0 - margin)
-        x2, y2 = self.to_local((self.width + margin),
-                               (self.height + margin))
+        x2, y2 = self.to_local((self.width + margin), (self.height + margin))
         c1 = self.get_latlon_at(x1, y1)
         c2 = self.get_latlon_at(x2, y2)
         return Bbox((c1.lat, c1.lon, c2.lat, c2.lon))
@@ -395,23 +409,25 @@ class MapView(Widget):
         """Sets the zoom level, leaving the (x, y) at the exact same point
         in the view.
         """
-        zoom = clamp(zoom,
-                     self.map_source.get_min_zoom(),
-                     self.map_source.get_max_zoom())
+        zoom = clamp(
+            zoom, self.map_source.get_min_zoom(), self.map_source.get_max_zoom()
+        )
         if int(zoom) == int(self._zoom):
             if scale is None:
                 return
             elif scale == self.scale:
                 return
-        scale = scale or 1.
+        scale = scale or 1.0
 
         # first, rescale the scatter
         scatter = self._scatter
         scale = clamp(scale, scatter.scale_min, scatter.scale_max)
         rescale = scale * 1.0 / scatter.scale
-        scatter.apply_transform(Matrix().scale(rescale, rescale, rescale),
-                                post_multiply=True,
-                                anchor=scatter.to_local(x, y))
+        scatter.apply_transform(
+            Matrix().scale(rescale, rescale, rescale),
+            post_multiply=True,
+            anchor=scatter.to_local(x, y),
+        )
 
         # adjust position if the zoom changed
         c1 = self.map_source.get_col_count(self._zoom)
@@ -421,9 +437,9 @@ class MapView(Widget):
             self.delta_x = scatter.x + self.delta_x * f
             self.delta_y = scatter.y + self.delta_y * f
             # back to 0 every time
-            scatter.apply_transform(Matrix().translate(
-                -scatter.x, -scatter.y, 0
-            ), post_multiply=True)
+            scatter.apply_transform(
+                Matrix().translate(-scatter.x, -scatter.y, 0), post_multiply=True
+            )
 
         # avoid triggering zoom changes.
         self._zoom = zoom
@@ -447,7 +463,8 @@ class MapView(Widget):
         scale = self._scale
         return Coordinate(
             lat=self.map_source.get_lat(zoom, y / scale + vy),
-            lon=self.map_source.get_lon(zoom, x / scale + vx))
+            lon=self.map_source.get_lon(zoom, x / scale + vx),
+        )
 
     def add_marker(self, marker, layer=None):
         """Add a marker into the layer. If layer is None, it will be added in
@@ -477,9 +494,8 @@ class MapView(Widget):
         widget yourself: think as Z-sprite / billboard.
         Defaults to "window".
         """
-        assert (mode in ("scatter", "window"))
-        if self._default_marker_layer is None and \
-                isinstance(layer, MarkerMapLayer):
+        assert mode in ("scatter", "window")
+        if self._default_marker_layer is None and isinstance(layer, MarkerMapLayer):
             self._default_marker_layer = layer
         self._layers.append(layer)
         c = self.canvas
@@ -488,7 +504,7 @@ class MapView(Widget):
         else:
             self.canvas = self.canvas_layers_out
         layer.canvas_parent = self.canvas
-        super(MapView, self).add_widget(layer)
+        super().add_widget(layer)
         self.canvas = c
 
     def remove_layer(self, layer):
@@ -497,7 +513,7 @@ class MapView(Widget):
         c = self.canvas
         self._layers.remove(layer)
         self.canvas = layer.canvas_parent
-        super(MapView, self).remove_widget(layer)
+        super().remove_widget(layer)
         self.canvas = c
 
     def sync_to(self, other):
@@ -511,6 +527,7 @@ class MapView(Widget):
 
     def __init__(self, **kwargs):
         from kivy.base import EventLoop
+
         EventLoop.ensure_window()
         self._invalid_scale = True
         self._tiles = []
@@ -530,13 +547,13 @@ class MapView(Widget):
         with self.canvas:
             self.canvas_layers_out = Canvas()
         self._scale_target_anim = False
-        self._scale_target = 1.
+        self._scale_target = 1.0
         self._touch_count = 0
         self.map_source.cache_dir = self.cache_dir
-        Clock.schedule_interval(self._animate_color, 1 / 60.)
+        Clock.schedule_interval(self._animate_color, 1 / 60.0)
         self.lat = kwargs.get("lat", self.lat)
         self.lon = kwargs.get("lon", self.lon)
-        super(MapView, self).__init__(**kwargs)
+        super().__init__(**kwargs)
 
     def _animate_color(self, dt):
         # fast path
@@ -544,14 +561,14 @@ class MapView(Widget):
         if d == 0:
             for tile in self._tiles:
                 if tile.state == "need-animation":
-                    tile.g_color.a = 1.
+                    tile.g_color.a = 1.0
                     tile.state = "animated"
             for tile in self._tiles_bg:
                 if tile.state == "need-animation":
-                    tile.g_color.a = 1.
+                    tile.g_color.a = 1.0
                     tile.state = "animated"
         else:
-            d = d / 1000.
+            d = d / 1000.0
             for tile in self._tiles:
                 if tile.state != "need-animation":
                     continue
@@ -571,7 +588,7 @@ class MapView(Widget):
         elif isinstance(widget, MapLayer):
             self.add_layer(widget)
         else:
-            super(MapView, self).add_widget(widget)
+            super().add_widget(widget)
 
     def remove_widget(self, widget):
         if isinstance(widget, MapMarker):
@@ -579,13 +596,13 @@ class MapView(Widget):
         elif isinstance(widget, MapLayer):
             self.remove_layer(widget)
         else:
-            super(MapView, self).remove_widget(widget)
+            super().remove_widget(widget)
 
     def on_map_relocated(self, zoom, coord):
         pass
 
     def animated_diff_scale_at(self, d, x, y):
-        self._scale_target_time = 1.
+        self._scale_target_time = 1.0
         self._scale_target_pos = x, y
         if self._scale_target_anim is False:
             self._scale_target_anim = True
@@ -593,10 +610,10 @@ class MapView(Widget):
         else:
             self._scale_target += d
         Clock.unschedule(self._animate_scale)
-        Clock.schedule_interval(self._animate_scale, 1 / 60.)
+        Clock.schedule_interval(self._animate_scale, 1 / 60.0)
 
     def _animate_scale(self, dt):
-        diff = self._scale_target / 3.
+        diff = self._scale_target / 3.0
         if abs(diff) < 0.01:
             diff = self._scale_target
             self._scale_target = 0
@@ -618,17 +635,18 @@ class MapView(Widget):
         scatter = self._scatter
         scale = clamp(scale, scatter.scale_min, scatter.scale_max)
         rescale = scale * 1.0 / scatter.scale
-        scatter.apply_transform(Matrix().scale(rescale, rescale, rescale),
-                                post_multiply=True,
-                                anchor=scatter.to_local(x, y))
+        scatter.apply_transform(
+            Matrix().scale(rescale, rescale, rescale),
+            post_multiply=True,
+            anchor=scatter.to_local(x, y),
+        )
 
     def on_touch_down(self, touch):
         if not self.collide_point(*touch.pos):
             return
         if self.pause_on_action:
             self._pause = True
-        if "button" in touch.profile and touch.button in (
-        "scrolldown", "scrollup"):
+        if "button" in touch.profile and touch.button in ("scrolldown", "scrollup"):
             d = 1 if touch.button == "scrollup" else -1
             self.animated_diff_scale_at(d, *touch.pos)
             return True
@@ -639,7 +657,7 @@ class MapView(Widget):
         self._touch_count += 1
         if self._touch_count == 1:
             self._touch_zoom = (self.zoom, self._scale)
-        return super(MapView, self).on_touch_down(touch)
+        return super().on_touch_down(touch)
 
     def on_touch_up(self, touch):
         if touch.grab_current == self:
@@ -651,12 +669,12 @@ class MapView(Widget):
                 cur_zoom = self.zoom
                 cur_scale = self._scale
                 if cur_zoom < zoom or cur_scale < scale:
-                    self.animated_diff_scale_at(1. - cur_scale, *touch.pos)
+                    self.animated_diff_scale_at(1.0 - cur_scale, *touch.pos)
                 elif cur_zoom > zoom or cur_scale > scale:
-                    self.animated_diff_scale_at(2. - cur_scale, *touch.pos)
+                    self.animated_diff_scale_at(2.0 - cur_scale, *touch.pos)
                 self._pause = False
             return True
-        return super(MapView, self).on_touch_up(touch)
+        return super().on_touch_up(touch)
 
     def on_transform(self, *args):
         self._invalid_scale = True
@@ -668,19 +686,19 @@ class MapView(Widget):
         zoom = self._zoom
         scatter = self._scatter
         scale = scatter.scale
-        if scale >= 2.:
+        if scale >= 2.0:
             zoom += 1
-            scale /= 2.
+            scale /= 2.0
         elif scale < 1:
             zoom -= 1
-            scale *= 2.
+            scale *= 2.0
         zoom = clamp(zoom, map_source.min_zoom, map_source.max_zoom)
         if zoom != self._zoom:
             self.set_zoom_at(zoom, scatter.x, scatter.y, scale=scale)
             self.trigger_update(True)
         else:
-            if zoom == map_source.min_zoom and scatter.scale < 1.:
-                scatter.scale = 1.
+            if zoom == map_source.min_zoom and scatter.scale < 1.0:
+                scatter.scale = 1.0
                 self.trigger_update(True)
             else:
                 self.trigger_update(False)
@@ -705,16 +723,16 @@ class MapView(Widget):
         oxmin, oymin = self._scatter.to_local(self.x, self.y)
         oxmax, oymax = self._scatter.to_local(self.right, self.top)
         s = self._scale
-        cxmin = (oxmin - dx)
+        cxmin = oxmin - dx
         if cxmin < xmin:
             self._scatter.x += (cxmin - xmin) * s
-        cymin = (oymin - dy)
+        cymin = oymin - dy
         if cymin < ymin:
             self._scatter.y += (cymin - ymin) * s
-        cxmax = (oxmax - dx)
+        cxmax = oxmax - dx
         if cxmax > xmax:
             self._scatter.x -= (xmax - cxmax) * s
-        cymax = (oymax - dy)
+        cymax = oymax - dy
         if cymax > ymax:
             self._scatter.y -= (ymax - cymax) * s
 
@@ -730,12 +748,12 @@ class MapView(Widget):
     def do_update(self, dt):
         zoom = self._zoom
         scale = self._scale
-        self.lon = self.map_source.get_lon(zoom,
-                                           (
-                                           self.center_x - self._scatter.x) / scale - self.delta_x)
-        self.lat = self.map_source.get_lat(zoom,
-                                           (
-                                           self.center_y - self._scatter.y) / scale - self.delta_y)
+        self.lon = self.map_source.get_lon(
+            zoom, (self.center_x - self._scatter.x) / scale - self.delta_x
+        )
+        self.lat = self.map_source.get_lat(
+            zoom, (self.center_y - self._scatter.y) / scale - self.delta_y
+        )
         self.dispatch("on_map_relocated", zoom, Coordinate(self.lon, self.lat))
         for layer in self._layers:
             layer.reposition()
@@ -768,8 +786,7 @@ class MapView(Widget):
 
         x_count = tile_x_last - tile_x_first
         y_count = tile_y_last - tile_y_first
-        return (tile_x_first, tile_y_first, tile_x_last, tile_y_last,
-                x_count, y_count)
+        return (tile_x_first, tile_y_first, tile_x_last, tile_y_last, x_count, y_count)
 
     def load_visible_tiles(self):
         map_source = self.map_source
@@ -779,12 +796,14 @@ class MapView(Widget):
         bbox_for_zoom = self.bbox_for_zoom
         size = map_source.dp_tile_size
 
-        tile_x_first, tile_y_first, tile_x_last, tile_y_last, \
-        x_count, y_count = bbox_for_zoom(vx, vy, self.width, self.height, zoom)
-
-        # print "Range {},{} to {},{}".format(
-        #    tile_x_first, tile_y_first,
-        #    tile_x_last, tile_y_last)
+        (
+            tile_x_first,
+            tile_y_first,
+            tile_x_last,
+            tile_y_last,
+            x_count,
+            y_count,
+        ) = bbox_for_zoom(vx, vy, self.width, self.height, zoom)
 
         # Adjust tiles behind us
         for tile in self._tiles_bg[:]:
@@ -794,11 +813,21 @@ class MapView(Widget):
             f = 2 ** (zoom - tile.zoom)
             w = self.width / f
             h = self.height / f
-            btile_x_first, btile_y_first, btile_x_last, btile_y_last, \
-            _, _ = bbox_for_zoom(vx / f, vy / f, w, h, tile.zoom)
+            (
+                btile_x_first,
+                btile_y_first,
+                btile_x_last,
+                btile_y_last,
+                _,
+                _,
+            ) = bbox_for_zoom(vx / f, vy / f, w, h, tile.zoom)
 
-            if tile_x < btile_x_first or tile_x >= btile_x_last or \
-                    tile_y < btile_y_first or tile_y >= btile_y_last:
+            if (
+                tile_x < btile_x_first
+                or tile_x >= btile_x_last
+                or tile_y < btile_y_first
+                or tile_y >= btile_y_last
+            ):
                 tile.state = "done"
                 self._tiles_bg.remove(tile)
                 self.canvas_map.before.remove(tile.g_color)
@@ -807,17 +836,19 @@ class MapView(Widget):
 
             tsize = size * f
             tile.size = tsize, tsize
-            tile.pos = (
-                tile_x * tsize + self.delta_x,
-                tile_y * tsize + self.delta_y)
+            tile.pos = (tile_x * tsize + self.delta_x, tile_y * tsize + self.delta_y)
 
         # Get rid of old tiles first
         for tile in self._tiles[:]:
             tile_x = tile.tile_x
             tile_y = tile.tile_y
 
-            if tile_x < tile_x_first or tile_x >= tile_x_last or \
-                    tile_y < tile_y_first or tile_y >= tile_y_last:
+            if (
+                tile_x < tile_x_first
+                or tile_x >= tile_x_last
+                or tile_y < tile_y_first
+                or tile_y >= tile_y_last
+            ):
                 tile.state = "done"
                 self.tile_map_set(tile_x, tile_y, False)
                 self._tiles.remove(tile)
@@ -825,8 +856,7 @@ class MapView(Widget):
                 self.canvas_map.remove(tile.g_color)
             else:
                 tile.size = (size, size)
-                tile.pos = (
-                tile_x * size + self.delta_x, tile_y * size + self.delta_y)
+                tile.pos = (tile_x * size + self.delta_x, tile_y * size + self.delta_y)
 
         # Load new tiles if needed
         x = tile_x_first + x_count // 2 - 1
@@ -836,9 +866,13 @@ class MapView(Widget):
         turn = 0
         while arm_size < arm_max:
             for i in range(arm_size):
-                if not self.tile_in_tile_map(x, y) and \
-                        y >= tile_y_first and y < tile_y_last and \
-                        x >= tile_x_first and x < tile_x_last:
+                if (
+                    not self.tile_in_tile_map(x, y)
+                    and y >= tile_y_first
+                    and y < tile_y_last
+                    and x >= tile_x_first
+                    and x < tile_x_last
+                ):
                     self.load_tile(x, y, size, zoom)
 
                 x += dirs[turn % 4 + 1]
@@ -852,7 +886,7 @@ class MapView(Widget):
     def load_tile(self, x, y, size, zoom):
         if self.tile_in_tile_map(x, y) or zoom != self._zoom:
             return
-        self.load_tile_for_source(self.map_source, 1., size, x, y, zoom)
+        self.load_tile_for_source(self.map_source, 1.0, size, x, y, zoom)
         # XXX do overlay support
         self.tile_map_set(x, y, True)
 
@@ -947,15 +981,19 @@ class MapView(Widget):
             self.map_source = MapSource.from_provider(source)
         elif isinstance(source, (tuple, list)):
             cache_key, min_zoom, max_zoom, url, attribution, options = source
-            self.map_source = MapSource(url=url, cache_key=cache_key,
-                                        min_zoom=min_zoom, max_zoom=max_zoom,
-                                        attribution=attribution,
-                                        cache_dir=self.cache_dir, **options)
+            self.map_source = MapSource(
+                url=url,
+                cache_key=cache_key,
+                min_zoom=min_zoom,
+                max_zoom=max_zoom,
+                attribution=attribution,
+                cache_dir=self.cache_dir,
+                **options
+            )
         elif isinstance(source, MapSource):
             self.map_source = source
         else:
             raise Exception("Invalid map source provider")
-        self.zoom = clamp(self.zoom,
-                          self.map_source.min_zoom, self.map_source.max_zoom)
+        self.zoom = clamp(self.zoom, self.map_source.min_zoom, self.map_source.max_zoom)
         self.remove_all_tiles()
         self.trigger_update(True)